+++ MapboxNavigationObjc/include/MapboxNavigation.h
@@ -7,4 +7,4 @@ FOUNDATION_EXPORT double MapboxNavigationVersionNumber;
 FOUNDATION_EXPORT const unsigned char MapboxNavigationVersionString[];
 
 #import "../MBRouteVoiceController.h"
-#import "../MGLMapView+MGLNavigationAdditions.h"
+#import "../MLNMapView+MLNNavigationAdditions.h"