+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changes to the Mapbox Navigation SDK for iOS
 
 ## Unreleased
+ - Uses the `Locale` given in `RouteOptions` to create the corresponding `AVSpeechSynthesisVoice`.
  - Removed setCamera() from updateCourseTracking()
  - Added setCamera() to progressDidChange()
  - Allow to customize rerouting logic by implementing RouteControllerDelegate#routeControllerGetDirections 
  - Add option to overwrite camera update via NavigationMapViewCourseTrackingDelegate#updateCamera
  - Remove MapboxVoiceController and Mapbox Speech dependency. If you would like to use MapboxSpeech, you can copy the deleted MapboxVoiceController into your project.
+ - Updated MapLibre Native dependency to ios-v6.0.0 (https://github.com/maplibre/maplibre-native/releases/tag/ios-v6.0.0). Implementers need to change the prefix MGL to MLN for all MapLibre Native classes that are referenced.
+ - Only snap location to route if the location is within the `RouteControllerUserLocationSnappingDistance`
  - Add support for Swift Package Manager while dropping Carthage and Cocoapods.
 
 ## v2.0.0 (May 23, 2023)